+++ vignettes/articles/DevGuide.R
@@ -188,11 +188,12 @@ Design(oms)
 rownames(Design(oms)) <- c('Base Case', 'Less Productive', 'More Productive')
 
 ## ---- om_preset ----
+# each element is a vector of row indices into `Design`
 
-Preset(oms) <- list('Base Case'=list(1),
-                    'Low M' = list(2),
-                    'High M' = list(3),
-                    'All'= list(1:3)
+Preset(oms) <- list('Base Case'=1,
+                    'Low M' = 2,
+                    'High M' = 3,
+                    'All'= 1:3
 )